% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivis.R
\name{ivis}
\alias{ivis}
\title{IVIS algorithm}
\usage{
ivis(X, y = NULL, embedding_dims = 2L, k = 150L, distance = "pn",
  batch_size = 128L, epochs = 1000L, n_epochs_without_progress = 50L,
  margin = 1, ntrees = 50L, search_k = -1L, precompute = TRUE)
}
\arguments{
\item{X}{numerical matrix to be reduced. Columns correspond to features.}

\item{y}{int class vector guiding supervised triplet selection.}

\item{embedding_dims}{int, optional (default: 2) Number of dimensions in the embedding space}

\item{distance}{string, optional (default: "pn")
The loss function used to train the neural network. One of "pn", "euclidean", "softmax_ratio_pn", "softmax_ratio".}

\item{batch_size}{int, optional (default: 128)
The size of mini-batches used during gradient descent while training the neural network.}

\item{epochs}{int, optional (default: 1000)
The maximum number of epochs to train the model for. Each epoch the network will see a triplet based on each data-point once.}

\item{n_epochs_without_progress}{int, optional (default: 50)
After n number of epochs without an improvement to the loss, terminate training early.}

\item{margin}{float, optional (default: 1)
The distance that is enforced between points by the triplet loss functions}

\description{
IVIS algorithm
}
